Remove Perl 4-ish function call syntax
Most of the functions were called as &Function; I don't know if I just
thought this was a good idea, having barely come out of Perl 4, or what.
Lost in time, like tears in the rain. Time for those ampersands to die.

Also needed to add () to the former &Function calls so that strict
wouldn't freak out about barewords.