Make fatal and note less complicated
They really just need to return the callback, and the code using them
just needs to call the callback as a function, like

   $obj->fatal->("arguments go here");

Tiny bit more load on the caller, but now these are just straight
setters and getters.

Also change note to use `carp` by default; if we really want to see what
was going on, it's nicer to have the caller's line number instead.

Also make sure that debug(), sleep() and tries() all work the same way.