Move old code to hold/; start Mojo app
Moved the original code that had been rewrapped as a module to hold/ so
we won't accidentally run it; we will pick and choose parts out of the
hold/ module and rewrite them as we go.

Added a new App::WebWebXNG module that loads Mojo and doesn't do
anything else yet.